fix payment + cart

+++ src/main/java/com/dat/backend/datshop/payment/service/impl/PaymentServiceImpl.java
@@ -12,7 +12,7 @@ import com.dat.backend.datshop.payment.entity.BillStatus;
 import com.dat.backend.datshop.payment.entity.Bill_Items;
 import com.dat.backend.datshop.payment.repository.BillRepository;
 import com.dat.backend.datshop.payment.repository.Bill_ItemsRepository;
-import com.dat.backend.datshop.payment.service.VNPayService;
+import com.dat.backend.datshop.payment.service.PaymentService;
 import com.dat.backend.datshop.payment.util.VNPayUtil;
 import com.dat.backend.datshop.product.entity.Product;
 import com.dat.backend.datshop.product.repository.ProductRepository;
@@ -21,14 +21,18 @@ import com.dat.backend.datshop.user.repository.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
-public class VNPayServiceImpl implements VNPayService {
+public class PaymentServiceImpl implements PaymentService {
     private final VNPayConfig vnPayConfig;
     private final UserRepository userRepository;
     private final BillRepository billRepository;
@@ -37,6 +41,7 @@ public class VNPayServiceImpl implements VNPayService {
     private final CartItemRepository cartItemRepository;
     private final Bill_ItemsRepository billItemsRepository;
 
+    // VNPay payment URL creation and bill creation
     @Override
     @Transactional
     public BillResponse createPayment(List<PayRequest> payRequestList, HttpServletRequest request, String email) {
@@ -50,19 +55,21 @@ public class VNPayServiceImpl implements VNPayService {
                 .mapToLong(payRequest -> {
                     Product product = productRepository.findById(payRequest.getProductId())
                             .orElseThrow(() -> new RuntimeException("Product not found with ID: " + payRequest.getProductId()));
+                    log.info("Product found: {}, Price: {}, Quantity: {}", product.getName(), product.getPrice(), payRequest.getQuantity());
                     return (long) (product.getPrice() * payRequest.getQuantity());
                 })
                 .sum();
 
+        log.info("Total amount to be paid: {}", amount);
+
         // Create bill entity
         Bill bill = Bill.builder()
+                .id(UUID.randomUUID().toString())
                 .userId(user.getId())
                 .price(amount)
                 .status(BillStatus.PENDING)
                 .description("Waiting for payment")
                 .build();
-        billRepository.save(bill);
-
 
         // Check stock
         for (PayRequest payRequest : payRequestList) {
@@ -82,16 +89,25 @@ public class VNPayServiceImpl implements VNPayService {
             billItemsRepository.save(billItem);
         }
 
+        // Convert amount to long for VNPay
+        long totalAmount = (long) (amount * 100); // Convert to VND
+
+        log.info("Total amount to be paid: {}", totalAmount);
+
         // Create payment URL
         String ipAddr = VNPayUtil.getIpAddr(request);
         Map<String, String> vnpParamsMap = vnPayConfig.createVnPayUrl(bill.getId());
-        vnpParamsMap.put("vnp_Amount", String.valueOf(amount*100L));
+        vnpParamsMap.put("vnp_Amount", String.valueOf(totalAmount));
         vnpParamsMap.put("vnp_IpAddr", ipAddr);
         String queryUrl = VNPayUtil.getPaymentURL(vnpParamsMap, true);
         String hashData = VNPayUtil.getPaymentURL(vnpParamsMap, false);
         String vnpSecureHash = VNPayUtil.hmacSHA512(vnPayConfig.vnp_HashSecret, hashData);
         queryUrl += "&vnp_SecureHash=" + vnpSecureHash;
         String paymentUrl = vnPayConfig.vnp_PayUrl + "?" + queryUrl;
+
+        // Save bill to database
+        bill.setPaymentUrl(paymentUrl);
+        billRepository.save(bill);
 
         // Create BillResponse
         return BillResponse.builder()
@@ -107,8 +123,7 @@ public class VNPayServiceImpl implements VNPayService {
     @Override
     public String paymentCallbackHandler(HttpServletRequest request) {
         String status = request.getParameter("vnp_ResponseCode");
-        String billIdStr = request.getParameter("billId");
-        Long billId = Long.parseLong(billIdStr);
+        String billId = request.getParameter("billId");
         if (status.equals("00")) {
             /*             * Payment successful
              * Update bill status to PAID and reduce stock quantity
@@ -118,8 +133,7 @@ public class VNPayServiceImpl implements VNPayService {
             Bill bill = billRepository.findById(billId)
                     .orElseThrow(() -> new RuntimeException("Bill not found with ID: " + billId));
 
-            Cart cart = cartRepository.findByUserId(bill.getUserId())
-                    .orElseThrow(() -> new RuntimeException("Cart not found for user ID: " + bill.getUserId()));
+            log.info("Bill found: {}", bill);
 
             // Remove bill items from the bill
             List<Bill_Items> billItems = billItemsRepository.findByBillId(billId);
@@ -135,16 +149,21 @@ public class VNPayServiceImpl implements VNPayService {
                 product.setStockQuantity(product.getStockQuantity() - billItem.getQuantity());
                 productRepository.save(product);
 
-                // Update cart item quantity
-                CartItem cartItem = cartItemRepository.findByCartIdAndProductId(cart.getId(), billItem.getProductId())
-                        .orElseThrow(() -> new RuntimeException("Cart item not found for product ID: " + billItem.getProductId()));
-                cartItem.setQuantity(cartItem.getQuantity() - billItem.getQuantity());
-                if (cartItem.getQuantity() <= 0) {
-                    // Remove cart item if quantity is zero or less
-                    cartItemRepository.delete(cartItem);
-                } else {
-                    // Save updated cart item
-                    cartItemRepository.save(cartItem);
+                // Find cart if user have added product to cart before
+                Cart cart = cartRepository.findByUserId(bill.getUserId())
+                        .orElseThrow(() -> new RuntimeException("Cart not found for user ID: " + bill.getUserId()));
+
+                Optional<CartItem> cartItemOpt = cartItemRepository.findByCartIdAndProductId(cart.getId(), billItem.getProductId());
+                if (cartItemOpt.isPresent()) {
+                    CartItem cartItem = cartItemOpt.get();
+                    cartItem.setQuantity(cartItem.getQuantity() - billItem.getQuantity());
+                    if (cartItem.getQuantity() <= 0) {
+                        // Remove cart item if quantity is zero or less
+                        cartItemRepository.delete(cartItem);
+                    } else {
+                        // Save updated cart item
+                        cartItemRepository.save(cartItem);
+                    }
                 }
 
                 // Remove bill item
@@ -161,5 +180,29 @@ public class VNPayServiceImpl implements VNPayService {
         } else {
             return "Failed";
         }
+    }
+
+    @Override
+    public List<BillResponse> getAllBills(String email) {
+        // Fetch user
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new RuntimeException("User not found with email: " + email));
+
+        // Fetch all bills for the user
+        List<Bill> bills = billRepository.findByUserId(user.getId());
+        if (bills.isEmpty()) {
+            throw new RuntimeException("No bills found for user with email: " + email);
+        }
+
+        // Map bills to BillResponse
+        return bills.stream()
+                .map(bill -> BillResponse.builder()
+                        .billId(bill.getId())
+                        .status(bill.getStatus().toString())
+                        .paymentUrl(bill.getPaymentUrl())
+                        .createdAt(bill.getCreatedAt())
+                        .updatedAt(bill.getUpdatedAt())
+                        .build())
+                .toList();
     }
 }
